Refactor de_export.py, extract _codegen.model_class
Move functionality for generating the ``amici::Model`` subclass code to a separate file.

Related to #2306
